requirements.txt
tensorflow 
keras 
numpy 
opencv-python 
pillow